Fixed sonar

+++ src/main/java/g3401_3500/s3464_maximize_the_distance_between_points_on_a_square/Solution.java
@@ -31,7 +31,7 @@ public class Solution {
             return x;
         }
         if (x == sideLength) {
-            return sideLength + (sideLength - y);
+            return sideLength + (long) (sideLength - y);
         }
         if (y == 0) {
             return 2L * sideLength + (sideLength - x);
